✨ : shared/ui 컴포넌트 세분화
- 기존 파일 내 각 컴포넌트 분리/세분화 작업 및 카테고라이징 작업
- 일부 컴포넌트 타입 작성 및 코드 정리
