+++ src/shared/ui/dialog/Dialog.tsx
@@ -1,0 +1,4 @@
+import * as DialogPrimitive from '@radix-ui/react-dialog';
+
+export const Dialog = DialogPrimitive.Root;
+export const DialogTrigger = DialogPrimitive.Trigger;
