+++ src/shared/ui/dialog/DialogHeader.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+
+export const DialogHeader: React.FC<React.HTMLAttributes<HTMLDivElement>> = ({
+  className,
+  ...props
+}) => (
+  <div className={`flex flex-col space-y-1.5 text-center sm:text-left ${className}`} {...props} />
+);
+
+DialogHeader.displayName = 'DialogHeader';
