+++ src/shared/ui/form/Textarea.tsx
@@ -1,0 +1,17 @@
+import { forwardRef, TextareaHTMLAttributes } from 'react';
+
+interface TextAreaProps extends TextareaHTMLAttributes<HTMLTextAreaElement> {
+  className?: string;
+}
+
+// 텍스트 영역 컴포넌트
+const baseTextareaStyles =
+  'flex min-h-[150px] w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50';
+
+export const Textarea = forwardRef<HTMLTextAreaElement, TextAreaProps>(
+  ({ className, ...props }, ref) => (
+    <textarea className={`${baseTextareaStyles} ${className}`} ref={ref} {...props} />
+  ),
+);
+
+Textarea.displayName = 'Textarea';
