+++ src/shared/ui/select/Select.tsx
@@ -1,0 +1,5 @@
+import * as SelectPrimitive from '@radix-ui/react-select';
+
+export const Select = SelectPrimitive.Root;
+export const SelectGroup = SelectPrimitive.Group;
+export const SelectValue = SelectPrimitive.Value;
