+++ src/shared/ui/table/TableBody.tsx
@@ -1,0 +1,10 @@
+import React, { forwardRef } from 'react';
+
+export const TableBody = forwardRef<
+  HTMLTableSectionElement,
+  React.HTMLAttributes<HTMLTableSectionElement>
+>(({ className, ...props }, ref) => (
+  <tbody ref={ref} className={`[&_tr:last-child]:border-0 ${className}`} {...props} />
+));
+
+TableBody.displayName = 'TableBody';
