+++ src/shared/ui/table/TableHeader.tsx
@@ -1,0 +1,10 @@
+import React, { forwardRef } from 'react';
+
+export const TableHeader = forwardRef<
+  HTMLTableSectionElement,
+  React.HTMLAttributes<HTMLTableSectionElement>
+>(({ className, ...props }, ref) => (
+  <thead ref={ref} className={`[&_tr]:border-b ${className}`} {...props} />
+));
+
+TableHeader.displayName = 'TableHeader';
