+++ src/shared/ui/table/TableRow.tsx
@@ -1,0 +1,13 @@
+import React, { forwardRef } from 'react';
+
+export const TableRow = forwardRef<HTMLTableRowElement, React.HTMLAttributes<HTMLTableRowElement>>(
+  ({ className, ...props }, ref) => (
+    <tr
+      ref={ref}
+      className={`border-b transition-colors hover:bg-muted/50 data-[state=selected]:bg-muted h-14 ${className}`}
+      {...props}
+    />
+  ),
+);
+
+TableRow.displayName = 'TableRow';
